+++ lib/src/pool_impl.dart
@@ -26,10 +26,12 @@ typedef Future<pg.Connection> ConnectionFactory(
      pg.TypeConverter? typeConverter,
      String? debugName});
 
-class ConnectionDecorator implements pg.Connection {
+class ConnectionDecorator implements pg.Connection, pgi.ConnectionOwner {
 
-  ConnectionDecorator(this._pool, PooledConnectionImpl pconn, this._conn)
-      : _pconn = pconn;
+  ConnectionDecorator(this._pool, PooledConnectionImpl pconn, pg.Connection conn)
+      : _pconn = pconn, _conn = conn {
+    if (conn is pgi.ConnectionImpl) conn.owner = this;
+  }
 
   _error(fnName) => new pg.PostgresqlException(
       '$fnName() called on closed connection.', _pconn.name);
@@ -41,8 +43,20 @@ class ConnectionDecorator implements pg.Connection {
 
   @override
   void close() {
-    if (!_isReleased) _pool._releaseConnection(_pconn);
-    _isReleased = true;
+    if (_release()) _pool._releaseConnection(_pconn);
+  }
+  @override
+  void destroy() {
+    if (_release()) _pool._destroyConnection(_pconn);
+  }
+
+  ///Returns false if it was released before.
+  bool _release() {
+    if (_isReleased) return false;
+
+    final conn = _conn;
+    if (conn is pgi.ConnectionImpl) conn.owner = _pconn; //restore it
+    return _isReleased = true;
   }
 
   @override
@@ -86,7 +100,7 @@ class ConnectionDecorator implements pg.Connection {
 }
 
 
-class PooledConnectionImpl implements PooledConnection {
+class PooledConnectionImpl implements PooledConnection, pgi.ConnectionOwner {
 
   PooledConnectionImpl(this._pool);
 
@@ -134,6 +148,10 @@ class PooledConnectionImpl implements PooledConnection {
       + (_useId == null ? '' : ':$_useId');
 
   @override
+  void destroy() {
+    _pool._destroyConnection(this);
+  }
+
   String toString() => '$name:$_state:$connectionState';
 }
 
@@ -260,6 +278,7 @@ class PoolImpl implements Pool {
         timeZone: settings.timeZone,
         typeConverter: _typeConverter,
         debugName: pconn.name);
+      if (conn is pgi.ConnectionImpl) conn.owner = pconn;
 
       // Pass this connection's messages through to the pool messages stream.
       conn.messages.listen((msg) => _messages.add(msg),
@@ -275,15 +294,25 @@ class PoolImpl implements Pool {
   }
 
   //A safe version that catches the exception.
-  Future _establishConnectionSafely()
-  => _establishConnection()
-    .catchError((ex) {
-      _messages.add(new pg.ClientMessage(
-          severity: 'WARNING',
-          message: "Failed to establish connection",
-          exception: ex));
-      return ex; //so caller can handle errors
-    });
+  Future _establishConnectionSafely() async {
+    for (DateTime since;;) //#21: try a while since PG likely recovering
+      try {
+        return _establishConnection();
+      } catch (ex) {
+        final now = DateTime.now();
+        if (since == null) {
+          since = now;
+          _messages.add(new pg.ClientMessage(
+              severity: 'WARNING',
+              message: "Failed to establish connection",
+              exception: ex));
+        } else if (now.difference(since) >= settings.connectionTimeout) {
+          return ex; //so caller can handle errors
+        }
+
+        await Future.delayed(const Duration(seconds: 1));
+      }
+  }
   
   void _heartbeat() {
     if (_state != running) return;
@@ -492,14 +521,14 @@ class PoolImpl implements Pool {
     //Purpose: make the caller of [connect] to end as soon as possible.
     //Otherwise, it will wait until timeout
     for (final r in results)
-      if (r is SocketException) { //unable to connect DB server
+      if (r != null) { //unable to connect DB server
         _processWaitQueue(); //dispatch succeeded conns first, if any
 
         final ex = new pg.PostgresqlException(
             'Failed to establish connection', null, exception: peConnectionFailed);
         while (_waitQueue.isNotEmpty)
           _waitQueue.removeAt(0).c.completeError(ex);
-        return; //done
+        break; //done
       }
   }
   bool _establishing = false;