+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 * `Settings.fromMap` and `Settings.toMap` are removed.
 * The `mockSocketConnect` argument is removed.
 
+### Version 0.7.8
+
+* Fix #20: remove connections from the pool when detecting server restarted/crashed
+* Fix #21: retry if failed to establish a connection
+
 ### Version 0.7.7
 
 * Rename `freeConnections` to `limitConnections`